requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
pyyaml>=6.0.0
jinja2>=3.1.0
